+++ app/javascript/lca/components/qcs/qcCharmFields.jsx
@@ -1,23 +1,30 @@
 // @flow
 import { isEqual } from 'lodash'
-import React, { Component } from 'react'
+import React from 'react'
+import { SortableHandle } from 'react-sortable-hoc'
 
 import Button from '@material-ui/core/Button'
-import Divider from '@material-ui/core/Divider'
+import Typography from '@material-ui/core/Typography'
 import Delete from '@material-ui/icons/Delete'
+import DragHandleIcon from '@material-ui/icons/DragHandle'
 
+import BlockPaper from 'components/generic/blockPaper.jsx'
 import CharmTimingSelect from 'components/generic/CharmTimingSelect'
 import RatingField from 'components/generic/RatingField.jsx'
 import TagsField from 'components/generic/TagsField.jsx'
 import TextField from 'components/generic/TextField.jsx'
 import type { QcCharm } from 'utils/flow-types'
 
+const Handle = SortableHandle(() => (
+  <DragHandleIcon onClick={e => e.preventDefault()} />
+))
+
 type Props = {
   charm: QcCharm,
   onCharmChange: Function,
   onRemoveClick: Function,
 }
-export default class QcCharmFields extends Component<Props> {
+export default class QcCharmFields extends React.Component<Props> {
   handleChange = (e: SyntheticInputEvent<>) => {
     const { name, value } = e.target
     const { charm } = this.props
@@ -36,7 +43,18 @@ export default class QcCharmFields extends Component<Props> {
     const { handleChange, handleRemove } = this
 
     return (
-      <div style={{ marginBottom: '0.75em' }}>
+      <BlockPaper>
+        <Typography
+          component="div"
+          style={{
+            position: 'absolute',
+            top: '0.5em',
+            right: '0.5em',
+            marginRight: '1em',
+          }}
+        >
+          <Handle />
+        </Typography>
         <TextField
           name="name"
           value={charm.name}
@@ -49,6 +67,7 @@ export default class QcCharmFields extends Component<Props> {
           value={charm.min_essence}
           label="Essence"
           margin="dense"
+          narrow
           onChange={handleChange}
         />
         <TextField
@@ -58,10 +77,6 @@ export default class QcCharmFields extends Component<Props> {
           margin="dense"
           onChange={handleChange}
         />
-        <Button onClick={handleRemove} style={{ float: 'right' }}>
-          Delete&nbsp;
-          <Delete />
-        </Button>
         <br />
         <CharmTimingSelect
           name="timing"
@@ -94,16 +109,26 @@ export default class QcCharmFields extends Component<Props> {
           multiline
           rowsMax={5}
         />
-        <br />
-        <TextField
-          name="ref"
-          value={charm.ref}
-          label="Reference"
-          margin="dense"
-          onChange={handleChange}
-        />
-        <Divider style={{ marginTop: '0.5em' }} />
-      </div>
+        <div style={{ display: 'flex', flexWrap: 'wrap' }}>
+          <TextField
+            name="ref"
+            value={charm.ref}
+            label="Reference"
+            margin="dense"
+            onChange={handleChange}
+            style={{ flex: 2 }}
+          />
+          <div style={{ flex: 1, textAlign: 'right' }}>
+            <Button
+              onClick={handleRemove}
+              style={{ margin: '1.5em -0.75em -1em 3em' }}
+            >
+              Delete&nbsp;
+              <Delete />
+            </Button>
+          </div>
+        </div>
+      </BlockPaper>
     )
   }
 }