+++ app/javascript/lca/components/qcs/qcMeritEditor.jsx
@@ -1,19 +1,25 @@
 // @flow
 import React from 'react'
 import { connect } from 'react-redux'
+import { SortableElement } from 'react-sortable-hoc'
 
 import Button from '@material-ui/core/Button'
+import Grid from '@material-ui/core/Grid'
 import Typography from '@material-ui/core/Typography'
 import ContentAddCircle from '@material-ui/icons/AddCircle'
 
 import QcMeritFields from './qcMeritFields.jsx'
+import SortableGridList from 'components/generic/SortableGridList.jsx'
 
 import { createQcMerit, destroyQcMerit, updateQcMerit } from 'ducks/actions.js'
 import { getMeritsForQc } from 'selectors'
 import type { fullQc, QcMerit, Enhancer } from 'utils/flow-types'
 
+const SortableItem = SortableElement(({ children }) => children)
+
 type ExposedProps = {
   qc: fullQc,
+  classes: Object,
 }
 type Props = ExposedProps & {
   qc_merits: Array<QcMerit>,
@@ -35,31 +41,53 @@ class QcMeritEditor extends React.Component<Props> {
     this.props.destroyQcMerit(id, this.props.qc.id)
   }
 
-  render() {
-    const { handleChange, handleAdd, handleRemove } = this
+  handleSort = ({ oldIndex, newIndex }) => {
+    if (oldIndex === newIndex) return
+    const meritA = this.props.qc_merits[oldIndex]
+    const meritB = this.props.qc_merits[newIndex]
+    const offset = meritA.sort_order > meritB.sort_order ? -1 : 1
+    this.props.updateQcMerit(
+      meritA.id,
+      this.props.qc.id,
+      'sort_order',
+      meritB.sort_order + offset
+    )
+  }
 
-    const qcMerits = this.props.qc_merits.map(merit => (
-      <QcMeritFields
-        key={merit.id}
-        merit={merit}
-        qc={this.props.qc}
-        onMeritChange={handleChange}
-        onRemoveClick={handleRemove}
-      />
+  render() {
+    const { handleChange, handleAdd, handleRemove, handleSort } = this
+    const { classes } = this.props
+
+    const qcMerits = this.props.qc_merits.map((merit, i) => (
+      <SortableItem key={merit.id} index={i}>
+        <Grid item xs={12} md={6} xl={4}>
+          <QcMeritFields
+            merit={merit}
+            qc={this.props.qc}
+            onMeritChange={handleChange}
+            onRemoveClick={handleRemove}
+          />
+        </Grid>
+      </SortableItem>
     ))
 
     return (
-      <div>
-        <Typography variant="title">
-          Merits
-          <Button onClick={handleAdd}>
-            Add Merit
-            <ContentAddCircle />
-          </Button>
-        </Typography>
-
-        {qcMerits}
-      </div>
+      <SortableGridList
+        header={
+          <Typography variant="title">
+            Merits
+            <Button onClick={handleAdd}>
+              Add Merit
+              <ContentAddCircle />
+            </Button>
+          </Typography>
+        }
+        items={qcMerits}
+        classes={classes}
+        onSortEnd={handleSort}
+        useDragHandle={true}
+        axis="xy"
+      />
     )
   }
 }